// whirlpool/plwhirlpool/Custom-apps/BE/WorldpayConnector/node/typings/WPCreatePayment/WPCreatePaymentRequest.d.ts
export interface WPCreatePaymentRequest {
  paymentService: {
    $: {
      version: string,
      merchantCode: string
    },
    submit: Order
  }
}
export interface WPCreateP24PaymentRequest extends WPCreatePaymentRequest {
  paymentService: {
    $: {
      version: string,
      merchantCode: string
    },
    submit: P24Order
  }
}

export interface Order {
  order:
  {
    $: {
      orderCode: string,
      installationId: string,
      captureDelay: string,
      shopperLanguageCode?: string,
    },
    description?: string,
    amount:
    {
      $: {
        currencyCode: string,
        exponent: string,
        value: string
      }
    },
    orderContent?: string,
    paymentMethodMask?: (Include | Exclude)[],
    paymentDetails?: (PRZELEWYSSL)[],
    shopper:
    {
      shopperEmailAddress: string
    },
    shippingAddress?: Address,
    billingAddress?: Address,
    createToken?: {
      $: {
        tokenScope: string
      },
      tokenEventReference: string,
      tokenReason: string
    }
  }
}

export interface P24Order extends Order {
  order:
  {
    $: {
      orderCode: string,
      shopperLanguageCode?: string,
      installationId: string,
      captureDelay: string
    },
    description?: string,
    amount:
    {
      $: {
        currencyCode: string,
        exponent: string,
        value: string
      }
    },
    orderContent?: string,
    paymentMethodMask?: (Include | Exclude)[],
    paymentDetails?: (PRZELEWYSSL)[],
    shopper:
    {
      shopperEmailAddress: string
    },
    shippingAddress?: Address,
    billingAddress?: Address,
    createToken?: {
      $: {
        tokenScope: string
      },
      tokenEventReference: string,
      tokenReason: string
    }
  }
}

export interface PRZELEWYSSL {
  "PRZELEWY-SSL": {
    $: {
      shopperCountryCode: string
    },
    successURL: string
    cancelURL: string
    pendingURL: string
  }
}

export interface Include {
  include:
  {
    $: {
      code: string
    }
  }
}

export interface Exclude {
  exclude:
  {
    $: {
      code: string
    }
  }
}

export interface Address {
  address:
  {
    address1: string,
    address2: string,
    address3?: string,
    postalCode: string,
    city: string,
    state?: string,
    countryCode: string
  }
}